requirements.txt
pytesseract
Pillow 
PyMuPDF # optional, for PDF/image handling 